backend/pyproject.toml
[project]
name = "demian-llm-journal-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.79.0",
    "fastapi>=0.129.0",
    "openai>=2.20.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "supabase>=2.28.0",
    "uvicorn>=0.40.0",
]
